Disable sql in post-auth
This reverts to not using sql for post-auth section, because it breaks
the authentication to freeradius.
The problem was observed when running ponsim tests, by comparing against
a working x86 deployment.
The conclusion was that the tpdock/freeradius:2.2.8 docker image
employed by the upstream charts, has been built almost 2 years ago,
meanwhile the sql option was enabled in 217ee9499, hence our
iecedge/freeradius:2.2.8 image differs in this aspect.

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>